Reducing Locking Impact of PlantHierarchy Refactoring Participants
For acquiring the plant hierarchy model the plant hierarchy refactoring
participants where locking the display not only for getting the plant
hierarchy view but also for loading it. In case of non existing plant
hierarchies this led to deadlocks. This fix reduces the lock time and
also only enables refactoring when an plant hierarchy file exists.