Fix inconsistent UI and validation for negated connections
It was possible to negate connections with an array of BOOL or InOut
variables at either end. Only the latter did yield a validation error.

This unifies the checks for negated connections by introducing a common
method in the model to check whether negation is supported. It also
allows to unnegate connections in the UI where negation is not
supported, so users are able to fix the error.

It further fixes live validation when negating connections by returning
the connection as affected object.